+++ tests/testthat/test_stream_tweets.R
@@ -1,17 +1,17 @@
 context("stream_tweets")
 
 test_that("stream_tweets returns tweets data", {
-	skip_on_cran()
+  skip_on_cran()
 
-	token <- readRDS("twitter_tokens")
-	x <- suppressMessages(stream_tweets("tweets", timeout = 2, verbose = FALSE,
-		token = token))
+  token <- readRDS("twitter_tokens")
+  x <- suppressMessages(stream_tweets("tweets", timeout = 2, verbose = FALSE,
+    token = token))
 
   expect_equal(is.data.frame(x), TRUE)
   expect_named(x)
   expect_true("status_id" %in% names(x))
   expect_gt(nrow(x), 0)
-  expect_gt(ncol(x), 23)
+  expect_gt(ncol(x), 15)
   expect_true("users" %in% names(attributes(x)))
   expect_true(is.data.frame(attr(x, "users")))
   expect_true(is.data.frame(users_data(x)))